feat: renamed to yahoo finance price
feat: requiring interval parameter
