+++ airbyte-integrations/connectors/source-yahoo-finance-price/.dockerignore
@@ -1,6 +1,6 @@
 *
 !Dockerfile
 !main.py
-!source_yahoo_finance
+!source_yahoo_finance_price
 !setup.py
 !secrets
