+++ airbyte-integrations/connectors/source-yahoo-finance-price/Dockerfile
@@ -29,10 +29,10 @@ RUN apk --no-cache add bash
 
 # copy payload code only
 COPY main.py ./
-COPY source_yahoo_finance ./source_yahoo_finance
+COPY source_yahoo_finance_price ./source_yahoo_finance_price
 
 ENV AIRBYTE_ENTRYPOINT "python /airbyte/integration_code/main.py"
 ENTRYPOINT ["python", "/airbyte/integration_code/main.py"]
 
 LABEL io.airbyte.version=0.1.0
-LABEL io.airbyte.name=airbyte/source-yahoo-finance
+LABEL io.airbyte.name=airbyte/source-yahoo-finance-price
