+++ airbyte-integrations/connectors/source-yahoo-finance-price/README.md
@@ -40,7 +40,7 @@ To build using Gradle, from the Airbyte repository root, run:
 
 #### Create credentials
 **If you are a community contributor**, follow the instructions in the [documentation](https://docs.airbyte.io/integrations/sources/yahoo-finance)
-to generate the necessary credentials. Then create a file `secrets/config.json` conforming to the `source_yahoo_finance/spec.json` file.
+to generate the necessary credentials. Then create a file `secrets/config.json` conforming to the `source_yahoo_finance_price/spec.json` file.
 Note that any directory named `secrets` is gitignored across the entire Airbyte repo, so there is no danger of accidentally checking in sensitive information.
 See `integration_tests/sample_config.json` for a sample config file.
