+++ airbyte-integrations/connectors/source-yahoo-finance-price/acceptance-test-config.yml
@@ -3,7 +3,7 @@
 connector_image: airbyte/source-yahoo-finance:dev
 tests:
   spec:
-    - spec_path: "source_yahoo_finance/spec.json"
+    - spec_path: "source_yahoo_finance_price/spec.json"
   connection:
     - config_path: "secrets/config.json"
       status: "succeed"
