+++ airbyte-integrations/connectors/source-yahoo-finance-price/acceptance-test-docker.sh
