+++ airbyte-integrations/connectors/source-yahoo-finance-price/integration_tests/abnormal_state.json
