+++ airbyte-integrations/connectors/source-yahoo-finance-price/integration_tests/catalog.json
