+++ airbyte-integrations/connectors/source-yahoo-finance-price/integration_tests/configured_catalog.json
