+++ airbyte-integrations/connectors/source-yahoo-finance-price/integration_tests/invalid_config.json
