+++ airbyte-integrations/connectors/source-yahoo-finance-price/integration_tests/sample_config.json
