+++ airbyte-integrations/connectors/source-yahoo-finance-price/source_yahoo_finance_price/schemas/price.json
