+++ airbyte-integrations/connectors/source-yahoo-finance-price/unit_tests/test_incremental_streams.py
@@ -5,7 +5,7 @@
 
 from airbyte_cdk.models import SyncMode
 from pytest import fixture
-from source_yahoo_finance.source import IncrementalYahooFinanceStream
+from source_yahoo_finance_price.source import IncrementalYahooFinanceStream
 
 
 @fixture
