--- airbyte-integrations/connectors/source-yahoo-finance/source_yahoo_finance/spec.json
@@ -1,19 +1,0 @@
-{
-  "documentationUrl": "https://docsurl.com",
-  "connectionSpecification": {
-    "$schema": "http://json-schema.org/draft-07/schema#",
-    "title": "Yahoo Finance Spec",
-    "type": "object",
-    "required": [
-      "ticker"
-    ],
-    "additionalProperties": false,
-    "properties": {
-      "ticker": {
-        "type": "string",
-        "description": "Identifier for the stock to be queried,",
-        "examples": ["AAPL, GOOGL, SPY"]
-      }
-    }
-  }
-}
